+++ tests/test_base_adapters.py
@@ -1,8 +1,8 @@
 import pytest
-from qcio import CalcType
+from qcio import CalcType, ProgramFailure
 
 from qcop.adapters import base, registry
-from qcop.exceptions import AdapterInputError
+from qcop.exceptions import AdapterInputError, ExternalProgramExecutionError
 
 
 def test_adapter_subclasses_must_define_program():
@@ -76,3 +76,37 @@ def test_adapters_raise_error_if_calctype_not_supported(prog_inp):
     gradient_input = prog_inp("gradient")
     with pytest.raises(AdapterInputError):
         TestAdapter().compute(gradient_input)
+
+
+def test_results_added_to_program_failure_object_if_exception_contains_them(
+    prog_inp, mocker, sp_output, test_adapter
+):
+    """Test that results are added to the ProgramFailure object if the exception
+    contains them."""
+    test_adapter = registry["test"]()
+
+    def raise_error(*args, **kwargs):
+        raise ExternalProgramExecutionError(
+            1,
+            "terachem tc.in",
+            stdout="some stdout",
+            results=sp_output.results,
+        )
+
+    mocker.patch.object(
+        test_adapter,
+        "compute_results",
+        side_effect=raise_error,
+    )
+    energy_input = prog_inp("energy")
+
+    # Check that the exception object contains the results
+    with pytest.raises(ExternalProgramExecutionError) as excinfo:
+        # import pdb; pdb.set_trace()
+        test_adapter.compute(energy_input, raise_exc=True)
+    assert excinfo.value.results == sp_output.results
+
+    # If no raise_exc=False, the results are added to the ProgramFailure
+    prog_failure = test_adapter.compute(energy_input)
+    assert isinstance(prog_failure, ProgramFailure)
+    assert prog_failure.results == sp_output.results